pyproject.toml
[project]
name = "eloApp"
version = "0.1.0"
description = "Welcoming workflow automation for middle-sized churches using Google Gemini as Data Structurer and streamlit as UI provider"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth-oauthlib>=1.2.2",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "streamlit-authenticator>=0.4.2",
]

[project.scripts]
eloapp = "eloApp.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

